README.md: prefer yarn over npm to install
Long story short: on my macOS system with node v8.1.4 and the
bundled npm (5.0.3), I had many troubles building the dependencies
and, even when that is done, I an error that `better-npm-run` is
missing.

I feel like we could use spend time really understanding what
is going on and making `npm` work, or we can just document the
way in which we develop, i.e. using `yarn`.

I took the second approach.